Update api.php to include charset option
In order to be able to connect to a mysql instance I had to change all instances of utf8mb4 to utf8. I just dies with a 9999 error without any hint as to why. As soon as I changed it to utf8 it worked for my database. So I figured there should be a way to override it so it is not set in stone. This is what this commit does.